+++ docs-demo/basic/row-cell-mouse-event/RowCellHoverSelect.vue
@@ -6,11 +6,14 @@ import CheckItem from '../../components/CheckItem.vue';
 const stkTableRef = useTemplateRef('stkTableRef');
 
 const stripe = ref(true);
-const rowActive = ref(true);
+const rowActive = ref({
+    enabled: true,
+    disabled: (row: any) => row.name === 'Disabled',
+    revokable: true,
+});
 const cellActive = ref(true);
 const rowHover = ref(true);
 const cellHover = ref(true);
-const rowCurrentRevokable = ref(true);
 const selectedCellRevokable = ref(true);
 
 const columns: StkTableColumn<any>[] = [
@@ -25,6 +28,7 @@ const dataSource = ref([
     { name: `Tom`, age: 20, address: `Shanghai`, gender: 'male' },
     { name: `Lucy`, age: 22, address: `Guangzhou`, gender: 'female' },
     { name: `Lily`, age: 24, address: `Shenzhen`, gender: 'female' },
+    { name: `Disabled`, age: 0, address: `Unknown`, gender: 'male' },
 ]);
 
 function setCurrentRow(rowKeyOrRow: string | undefined | any) {
@@ -38,15 +42,15 @@ function setSelectedCell(row: any, col: StkTableColumn<any>) {
 <template>
     <CheckItem v-model="stripe" text="stripe(斑马纹)"></CheckItem>
     <br />
-    <CheckItem v-model="rowActive" text="rowActive(行选中状态)"></CheckItem>
+    <CheckItem v-model="rowActive.enabled" text="rowActive(行选中状态)"></CheckItem>
     <CheckItem v-model="cellActive" text="cellActive(单元格选中状态)"></CheckItem>
     <br />
     <CheckItem v-model="rowHover" text="rowHover(行悬浮状态)"></CheckItem>
     <CheckItem v-model="cellHover" text="cellHover(单元格悬浮状态)"></CheckItem>
     <br />
     <CheckItem
-        v-model="rowCurrentRevokable"
-        text="rowCurrentRevokable(行选中状态是否可取消)"
+        v-model="rowActive.revokable"
+        text="rowActive.revokable(行选中状态是否可取消)"
     ></CheckItem>
     <br />
     <CheckItem
@@ -58,6 +62,7 @@ function setSelectedCell(row: any, col: StkTableColumn<any>) {
     <button class="btn" @click="setSelectedCell(dataSource[0], columns[1])">
         setSelectedCell('Jack-age')
     </button>
+    <button class="btn" @click="setCurrentRow('Disabled')">setCurrentRow('Disabled')</button>
 
     <StkTable
         ref="stkTableRef"
@@ -67,7 +72,6 @@ function setSelectedCell(row: any, col: StkTableColumn<any>) {
         :cell-active="cellActive"
         :row-hover="rowHover"
         :cell-hover="cellHover"
-        :row-current-revokable="rowCurrentRevokable"
         :selected-cell-revokable="selectedCellRevokable"
         :columns="columns"
         :data-source="dataSource"